seperate typenaming
