+++ src/testing/schema.spec.ts
@@ -12,7 +12,7 @@ import * as schemaDraft07 from "../schema/draft-07/index.js";
 import * as schema201909 from "../schema/draft-2019-09/index.js";
 import * as schema202012 from "../schema/draft-2020-12/index.js";
 import { SchemaContext } from "../schema/index.js";
-import { Namer, projectRoot } from "../utils/index.js";
+import { Namer, getNodeTypeName, projectRoot } from "../utils/index.js";
 
 const packageNames = [
     "string-or-boolean",
@@ -139,33 +139,4 @@ async function runTest(schemaName: string, packageName: string) {
             }
         });
     }
-}
-
-function getNodeTypeName(nodeUrl: URL): string {
-    const reReplace = /[^A-Za-z0-9-_.,]/gu;
-
-    const pointer = nodeUrl.hash.startsWith("#") ? nodeUrl.hash.substring(1) : "";
-
-    const pathParts = nodeUrl.pathname
-        .split("/")
-        .map(decodeURI)
-        .map((value) => value.replace(reReplace, ""))
-        .filter((value) => value !== "");
-    const pointerParts = pointer
-        .split("/")
-        .map(decodeURI)
-        .map((value) => value.replace(reReplace, ""));
-
-    const nameParts = [
-        pathParts[pathParts.length - 1] ?? "Schema",
-        pointerParts[pointerParts.length - 3],
-        pointerParts[pointerParts.length - 2],
-        pointerParts[pointerParts.length - 1],
-    ]
-        .filter((value) => value != null)
-        .filter((value) => value != "");
-
-    const name = camelcase(nameParts, { pascalCase: true });
-
-    return name;
 }
