+++ src/utils/index.ts
@@ -3,3 +3,4 @@ export * from "./literal.js";
 export * from "./namer.js";
 export * from "./package.js";
 export * from "./root.js";
+export * from "./type-name.js";
